requirements.txt
numpy
pillow
pytest
PyYAML
scipy
allure-pytest
pandas
serial
pytest-mock
